Rename travel-search-filters folder to map
