+++ src/map/storage.ts
